COMITTO

+++ WebApplication1/WebApplication1/Controllers/UsersController.cs
@@ -85,7 +85,7 @@ namespace WebApplication1.Controllers
 
         // DELETE: api/Users/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<User>> DeleteUser(long id)
+        public async Task<ActionResult<User>> DeleteUser(int id)
         {
             var user = await _context.Users.FindAsync(id);
             if (user == null)
